feat: extract @seans-mfe/oclif-base package (C3)
Moves BaseCommand and json-output helpers into packages/oclif-base/,
the package plugin authors depend on to get the JSON envelope, exit
codes, correlation ID, and prompt-rejection for free.  Old src/oclif/
BaseCommand.ts and json-output.ts become re-export shims.  Root CLI
and Jest module mapper updated with @seans-mfe/oclif-base alias.

Closes #111
