+++ package.json
@@ -72,6 +72,7 @@
   },
   "dependencies": {
     "@seans-mfe/contracts": "file:packages/contracts",
+    "@seans-mfe/oclif-base": "file:packages/oclif-base",
     "@apidevtools/swagger-parser": "^10.1.0",
     "@babel/generator": "^7.22.1",
     "@babel/parser": "^7.23.0",
