+++ packages/oclif-base/src/index.ts
@@ -1,0 +1,7 @@
+export { BaseCommand } from './BaseCommand';
+export {
+  suppressChalk,
+  redirectStdoutToStderr,
+  blockInteractivePrompts,
+  writeJsonLine,
+} from './json-output';
