+++ tsconfig.json
@@ -17,7 +17,9 @@
     "types": ["node", "jest"],
     "paths": {
       "@seans-mfe/contracts": ["./packages/contracts/src/index.ts"],
-      "@seans-mfe/contracts/*": ["./packages/contracts/src/*"]
+      "@seans-mfe/contracts/*": ["./packages/contracts/src/*"],
+      "@seans-mfe/oclif-base": ["./packages/oclif-base/src/index.ts"],
+      "@seans-mfe/oclif-base/*": ["./packages/oclif-base/src/*"]
     }
   },
   "include": ["src/**/*"],
